Simplifiy define-markup-[list-]command-internal,
This reverts additions made in commit
d2199b0163c33bcb7504c87e57eefbea93e08c88 "Issue 5167/3: Split off
`markup-lambda' from `define-markup-command'"

In this commit, extra support for the case where command-and-args is
empty was added, ie.

  -  (let* ((command (car command-and-args))
  -         (args (cdr command-and-args))
  +  (let* ((command (if (pair? command-and-args)
  +                      (car command-and-args)
  +                      command-and-args))
  +         (args (and (pair? command-and-args) (cdr command-and-args)))

However, markup commands are functions that are always called with
arguments 'layout' and 'props', so there can never be a case that
'args' would be empty.